ml/requirements.txt
# ML Pipeline Dependencies
# Install with: pip install -r requirements.txt

# Core data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Machine learning frameworks
scikit-learn>=1.3.0

# Gradient boosting models
lightgbm>=4.0.0
catboost>=1.2.0
xgboost>=2.0.0

# Hyperparameter optimization
optuna>=3.3.0

# Model persistence
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Database connectivity
psycopg2-binary>=2.9.0

# Environment management
python-dotenv>=1.0.0

# Optional: Progress bars
tqdm>=4.65.0

# API Server
flask>=3.0.0
flask-cors>=4.0.0
